# Enable Frozen string literals