--- a/lib/money.rb
+++ b/lib/money.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "bigdecimal"
 require "bigdecimal/util"
 require "set"