--- a/lib/money/bank/base.rb
+++ b/lib/money/bank/base.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   # Provides classes that aid in the ability of exchange one currency with
   # another.