--- a/lib/money/bank/single_currency.rb
+++ b/lib/money/bank/single_currency.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'money/bank/base'
 
 class Money