--- a/lib/money/locale_backend/base.rb
+++ b/lib/money/locale_backend/base.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'money/locale_backend/errors'
 
 class Money