--- a/lib/money/locale_backend/errors.rb
+++ b/lib/money/locale_backend/errors.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   module LocaleBackend
     class NotSupported < StandardError; end