--- a/lib/money/money.rb
+++ b/lib/money/money.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "money/bank/variable_exchange"
 require "money/bank/single_currency"
 require "money/money/arithmetic"