--- a/lib/money/money/allocation.rb
+++ b/lib/money/money/allocation.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   class Allocation
     # Splits a given amount in parts. The allocation is based on the parts' proportions