--- a/lib/money/money/arithmetic.rb
+++ b/lib/money/money/arithmetic.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   module Arithmetic
     # Wrapper for coerced numeric values to distinguish