--- a/lib/money/money/constructors.rb
+++ b/lib/money/money/constructors.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   module Constructors