--- a/lib/money/money/formatter.rb
+++ b/lib/money/money/formatter.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'money/money/formatting_rules'
 
 class Money
@@ -315,17 +317,19 @@ class Money
 
     def append_currency_symbol(formatted_number)
       if rules[:with_currency]
-        formatted_number << " "
+        currency_part =
+          if rules[:html]
+            "<span class=\"currency\">#{currency}</span>"
+          elsif rules[:html_wrap]
+            html_wrap(currency.to_s, "currency")
+          else
+            currency.to_s
+          end
 
-        if rules[:html]
-          formatted_number << "<span class=\"currency\">#{currency.to_s}</span>"
-        elsif rules[:html_wrap]
-          formatted_number << html_wrap(currency.to_s, "currency")
-        else
-          formatted_number << currency.to_s
-        end
+        "#{formatted_number} #{currency_part}"
+      else
+        formatted_number
       end
-      formatted_number
     end
 
     def show_free_text?