--- a/lib/money/money/formatting_rules.rb
+++ b/lib/money/money/formatting_rules.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   class FormattingRules
     def initialize(currency, *raw_rules)