--- a/lib/money/rates_store/memory.rb
+++ b/lib/money/rates_store/memory.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'monitor'
 
 class Money