--- a/lib/money/version.rb
+++ b/lib/money/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Money
   VERSION = '6.19.0'.freeze
 end