--- a/spec/bank/single_currency_spec.rb
+++ b/spec/bank/single_currency_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::Bank::SingleCurrency do
   describe "#exchange_with" do
     it "raises when called" do