--- a/spec/currency/heuristics_spec.rb
+++ b/spec/currency/heuristics_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::Currency::Heuristics do
   describe "#analyze_string" do
     it "it raises deprecation error" do