--- a/spec/currency/loader_spec.rb
+++ b/spec/currency/loader_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::Currency::Loader do
   it "returns a currency table hash" do
     expect(subject.load_currencies).to be_a Hash