--- a/spec/locale_backend/i18n_spec.rb
+++ b/spec/locale_backend/i18n_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::LocaleBackend::I18n do
   describe '#initialize' do
     it 'raises an error when I18n is not defined' do