--- a/spec/locale_backend/legacy_spec.rb
+++ b/spec/locale_backend/legacy_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::LocaleBackend::Legacy do
   after { Money.use_i18n = true }