--- a/spec/money/allocation_spec.rb
+++ b/spec/money/allocation_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::Allocation do
    describe 'given number as argument' do
      it 'raises an error when invalid argument is given' do