--- a/spec/money/formatting_rules_spec.rb
+++ b/spec/money/formatting_rules_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::FormattingRules do
   it 'does not modify frozen rules in place' do
     expect {