--- a/spec/money/locale_backend_spec.rb
+++ b/spec/money/locale_backend_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Money::LocaleBackend do
   describe '.find' do
     it 'returns an initialized backend' do