The only method that may potentially be affected is
`Money::Formatter#append_currency_symbol`.

This method has been refactored to avoid any potential side effect
with string mutations

Additionally, a `Lint/RedundantStringCoercion` offense has been fixed

Close #1108