@yukideluxe 2025-07-09
@tagliala, thanks for the PR and the explanation! I've double checked and, indeed, the only method that could be affected is `Money::Formatter#append_currency_symbol`. That said, I don't like the fact that it works because we know the `append_sign` returns an interpolated string and that is never frozen ([TIL](https://dictionary.cambridge.org/dictionary/english/til), btw 😳). 

I wonder if we should prevent unexpected consequences (eg. we want to use `Money::Formatter#append_currency_symbol` for other means in the future) and just rewrite that method to avoid the mutation. What do you think!?

Something like

```
    def append_currency_symbol(formatted_number)
      if rules[:with_currency]
        currency_part = 
          if rules[:html]
            "<span class=\"currency\">#{currency.to_s}</span>"
          elsif rules[:html_wrap]
            html_wrap(currency.to_s, "currency")
          else
            currency.to_s
          end

        "#{formatted_number} #{currency_part}"
      else
        formatted_number
      end
    end
```