@tjschuck 2025-07-09
The `frozen_string_literal` magic comment was introduced back in the Ruby 2.3 days, and a lot of stuff around strings has meaningfully changed since then.  So I'm not entirely sure this is _necessary_, but it doesn't per-se seem _harmful_.

Anyway, I benchmarked `Money#format` with/without this change, since https://github.com/RubyMoney/money/issues/1108 says that was the method that got the greatest improvement.  It seems... _fine_, and again, not particularly harmful or offensive.  I've gotten pretty good at ignoring these comments at the tops of files if they're there ¯\\\_(ツ)_/¯

First, the benchmark script, using `benchmark-ips` and `benchmark-memory`:

```ruby
# Run twice -- first do
#   $ ruby bench.rb
# then do
#   $ WITH_FROZEN_COMMENT=true ruby bench.rb

require 'bundler/inline'

if ENV['WITH_FROZEN_COMMENT'] == 'true'
  gemfile do
    source "https://rubygems.org"
    gem "benchmark-ips"
    gem "benchmark-memory"

    gem "money", git: 'https://github.com/tagliala/money', branch: 'feature/frozen-string-literals'
  end
else
  gemfile do
    source "https://rubygems.org"
    gem "benchmark-ips"
    gem "benchmark-memory"

    gem "money", git: 'https://github.com/RubyMoney/money', branch: 'main'
  end
end

require 'benchmark/ips'
require 'benchmark/memory'

I18n.config.available_locales = :en
Money.locale_backend = :i18n
Money.rounding_mode = BigDecimal::ROUND_HALF_UP

m = Money.from_cents(123, "GBP")

Benchmark.ips do |x|
  x.report('with frozen string comment') do
    m.format
  end

  x.report('without frozen string comment') do
    m.format
  end

  x.hold! 'temp_results_ips'
  x.compare!
end

Benchmark.memory do |x|
  x.report('with frozen string comment') do
    m.format
  end

  x.report('without frozen string comment') do
    m.format
  end

  x.hold! 'temp_results_mem'
  x.compare!
end
```

I ran the above on all supported Rubies (3.1 – 3.4), plus 3.5.0-preview1 for good measure.

### 3.1

```
Comparison:
with frozen string comment:    70473.7 i/s
without frozen string comment:    69933.5 i/s - same-ish: difference falls within error

Comparison:
with frozen string comment:       5296 allocated
without frozen string comment:       5776 allocated - 1.09x more
```

### 3.2

```
Comparison:
with frozen string comment:    80144.5 i/s
without frozen string comment:    77333.0 i/s - same-ish: difference falls within error

Comparison:
with frozen string comment:       5200 allocated
without frozen string comment:       5720 allocated - 1.10x more
```

### 3.3

```
Comparison:
with frozen string comment:    81450.4 i/s
without frozen string comment:    75517.5 i/s - same-ish: difference falls within error

Comparison:
with frozen string comment:       5176 allocated
without frozen string comment:       5696 allocated - 1.10x more
```

### 3.4

```
Comparison:
with frozen string comment:    77199.8 i/s
without frozen string comment:    74524.5 i/s - same-ish: difference falls within error

Comparison:
with frozen string comment:       4144 allocated
without frozen string comment:       4664 allocated - 1.13x more
```

### 3.5.0-preview1

```
Comparison:
with frozen string comment:    78967.5 i/s
without frozen string comment:    74296.0 i/s - same-ish: difference falls within error

Comparison:
with frozen string comment:       4144 allocated
without frozen string comment:       4664 allocated - 1.13x more
```

---

So... sure, go for it!