@tjschuck 2025-07-09
And yeah, let's rip out Code Climate next.  Fine to ignore the ❌ from it here.