@tagliala 2025-07-10
> So I'm not entirely sure this is necessary, but it doesn't per-se seem harmful.

The magic comment is still necessary because, as of now, no supported Ruby versions have frozen string literals enabled by default for all files.

Freezing string literals by default was originally planned for Ruby 3.0, but that change was postponed and has not happened yet. It may become the default in Ruby 4.0, but until then, the comment is needed

It is being discussed here: https://bugs.ruby-lang.org/issues/20205